+++ src/uk/co/ben_gibson/git/link/Url/Factory/BitbucketCloudUrlFactory.java
@@ -4,11 +4,12 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import uk.co.ben_gibson.git.link.Git.*;
 import uk.co.ben_gibson.git.link.Git.Exception.RemoteException;
+import uk.co.ben_gibson.git.link.UI.LineSelection;
 import uk.co.ben_gibson.git.link.Url.Factory.Exception.UrlFactoryException;
 
 import java.net.URL;
 
-public class BitBucketUrlFactory extends AbstractUrlFactory
+public class BitbucketCloudUrlFactory extends AbstractUrlFactory
 {
     public URL createUrlToCommit(@NotNull Remote remote, @NotNull Commit commit) throws UrlFactoryException, RemoteException
     {
@@ -22,15 +23,15 @@ public class BitBucketUrlFactory extends AbstractUrlFactory
         @NotNull Remote remote,
         @NotNull File file,
         @NotNull Branch branch,
-        @Nullable Integer lineNumber
+        @Nullable LineSelection lineSelection
     ) throws UrlFactoryException, RemoteException
     {
         String path = String.format("/%s/src/HEAD/%s", this.cleanPath(remote.url().getPath()), this.cleanPath(file.path()));
         String query = String.format("at=%s", branch.toString());
         String fragment = null;
 
-        if (lineNumber != null) {
-            fragment = String.format("%s-%d", file.name(), lineNumber);
+        if (lineSelection != null) {
+            fragment = this.formatLineSelection(lineSelection);
         }
 
         return this.buildURL(remote, path, query, fragment);
@@ -38,14 +39,19 @@ public class BitBucketUrlFactory extends AbstractUrlFactory
 
 
     @Override
-    public URL createUrlToFileAtCommit(@NotNull Remote remote, @NotNull File file, @NotNull Commit commit, @Nullable Integer lineNumber) throws UrlFactoryException, RemoteException
+    public URL createUrlToFileAtCommit(
+        @NotNull Remote remote,
+        @NotNull File file,
+        @NotNull Commit commit,
+        @Nullable LineSelection lineSelection
+    ) throws UrlFactoryException, RemoteException
     {
         String path = String.format("/%s/src/%s/%s", this.cleanPath(remote.url().getPath()), commit.hash(), this.cleanPath(file.path()));
 
         String fragment = null;
 
-        if (lineNumber != null) {
-            fragment = String.format("%s-%s", file.name(), lineNumber.toString());
+        if (lineSelection != null) {
+            fragment = this.formatLineSelection(lineSelection);
         }
 
         return this.buildURL(remote, path, null, fragment);
@@ -54,13 +60,16 @@ public class BitBucketUrlFactory extends AbstractUrlFactory
 
     public boolean supports(RemoteHost host)
     {
-        return host.isBitbucket();
+        return host.isBitbucketCloud();
     }
 
 
-    @Override
-    public boolean canOpenFileAtCommit()
+    private String formatLineSelection(LineSelection lineSelection)
     {
-        return true;
+        if (lineSelection.isMultiLineSelection()) {
+            return String.format("lines-%d:%d", lineSelection.start(), lineSelection.end());
+        }
+
+        return String.format("lines-%d", lineSelection.start());
     }
 }