+++ src/uk/co/ben_gibson/git/link/Url/Factory/GogsUrlFactory.java
@@ -4,6 +4,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import uk.co.ben_gibson.git.link.Git.*;
 import uk.co.ben_gibson.git.link.Git.Exception.RemoteException;
+import uk.co.ben_gibson.git.link.UI.LineSelection;
 import uk.co.ben_gibson.git.link.Url.Factory.Exception.UrlFactoryException;
 
 import java.net.URL;
@@ -19,18 +20,18 @@ public class GogsUrlFactory extends AbstractUrlFactory
 
 
     public URL createUrlToFileOnBranch(
-            @NotNull Remote remote,
-            @NotNull File file,
-            @NotNull Branch branch,
-            @Nullable Integer lineNumber
+        @NotNull Remote remote,
+        @NotNull File file,
+        @NotNull Branch branch,
+        @Nullable LineSelection lineSelection
     ) throws UrlFactoryException, RemoteException
     {
 
         String path = String.format("/%s/src/%s/%s", this.cleanPath(remote.url().getPath()), branch.toString(), this.cleanPath(file.path()));
         String fragment = null;
 
-        if (lineNumber != null) {
-            fragment = String.format("L%d", lineNumber);
+        if (lineSelection != null) {
+            fragment = this.formatLineSelection(lineSelection);
         }
 
         return this.buildURL(remote, path, null, fragment);
@@ -38,13 +39,18 @@ public class GogsUrlFactory extends AbstractUrlFactory
 
 
     @Override
-    public URL createUrlToFileAtCommit(@NotNull Remote remote, @NotNull File file, @NotNull Commit commit, @Nullable Integer lineNumber) throws UrlFactoryException, RemoteException
+    public URL createUrlToFileAtCommit(
+        @NotNull Remote remote,
+        @NotNull File file,
+        @NotNull Commit commit,
+        @Nullable LineSelection lineSelection
+    ) throws UrlFactoryException, RemoteException
     {
         String path = String.format("/%s/src/%s/%s", this.cleanPath(remote.url().getPath()), commit.hash(), this.cleanPath(file.path()));
         String fragment = null;
 
-        if (lineNumber != null) {
-            fragment = String.format("L%d", lineNumber);
+        if (lineSelection != null) {
+            fragment = this.formatLineSelection(lineSelection);
         }
 
         return this.buildURL(remote, path, null, fragment);
@@ -57,9 +63,12 @@ public class GogsUrlFactory extends AbstractUrlFactory
     }
 
 
-    @Override
-    public boolean canOpenFileAtCommit()
+    private String formatLineSelection(LineSelection lineSelection)
     {
-        return true;
+        if (lineSelection.isMultiLineSelection()) {
+            return String.format("L%d-L%d", lineSelection.start(), lineSelection.end());
+        }
+
+        return String.format("L%d", lineSelection.start());
     }
 }