+++ tests/unit/uk/co/ben_gibson/git/link/test/Url/Factory/GitBlitUrlFactoryTest.java
@@ -1,24 +1,17 @@
 package uk.co.ben_gibson.git.link.test.Url.Factory;
 
 import com.tngtech.java.junit.dataprovider.DataProvider;
-import org.junit.Test;
 import uk.co.ben_gibson.git.link.Git.Branch;
 import uk.co.ben_gibson.git.link.Git.Commit;
 import uk.co.ben_gibson.git.link.Git.Exception.RemoteException;
 import uk.co.ben_gibson.git.link.Git.File;
+import uk.co.ben_gibson.git.link.UI.LineSelection;
 import uk.co.ben_gibson.git.link.Url.Factory.GitBlitUrlFactory;
 
 import java.net.MalformedURLException;
 
 public class GitBlitUrlFactoryTest extends UrlFactoryTest
 {
-    @Test
-    public void testCanDetermineIfFileAtCommitIsSupported()
-    {
-        assertTrue(new GitBlitUrlFactory().canOpenFileAtCommit());
-    }
-
-
     @DataProvider
     public static Object[][] commitProvider() throws MalformedURLException, RemoteException
     {
@@ -61,7 +54,7 @@ public class GitBlitUrlFactoryTest extends UrlFactoryTest
                 mockRemote("http://my.company.com/gitblit/git/category/my.complex.repo"),
                 new File("src/com/foo/FooTest.java", "FooTest.java"),
                 new Commit("bd7ab0f04151e7409d77caa296617f97352f36d3"),
-                32,
+                new LineSelection(32),
                 "http://my.company.com/gitblit/blob/category!my.complex.repo.git/bd7ab0f04151e7409d77caa296617f97352f36d3/src!com!foo!FooTest.java#L32",
             },
         };
@@ -77,7 +70,7 @@ public class GitBlitUrlFactoryTest extends UrlFactoryTest
                 mockRemote("http://my.company.com/gitblit/git/category/my.complex.repo"),
                 new File("src/com/foo/FooTest.java", "FooTest.java"),
                 new Branch("feature/feature42D"),
-                32,
+                new LineSelection(32),
                 "http://my.company.com/gitblit/blob/category!my.complex.repo.git/feature!feature42D/src!com!foo!FooTest.java#L32",
             },
             {
@@ -85,7 +78,7 @@ public class GitBlitUrlFactoryTest extends UrlFactoryTest
                 mockRemote("http://my.company.com/git/r/ljpapi"),
                 new File("src/com/company/ljp/Alignment.java", "Alignment.java"),
                 new Branch("feature/feature42D"),
-                32,
+                new LineSelection(32),
                 "http://my.company.com/git/blob/ljpapi.git/feature!feature42D/src!com!company!ljp!Alignment.java#L32"
 
             },
@@ -94,7 +87,7 @@ public class GitBlitUrlFactoryTest extends UrlFactoryTest
                 mockRemote("http://my.company.com/r/ljpapi"),
                 new File("src/com/company/ljp/Alignment.java", "Alignment.java"),
                 new Branch("feature/feature42D"),
-                32,
+                new LineSelection(32),
                 "http://my.company.com/blob/ljpapi.git/feature!feature42D/src!com!company!ljp!Alignment.java#L32"
 
             },
@@ -103,7 +96,7 @@ public class GitBlitUrlFactoryTest extends UrlFactoryTest
                 mockRemote("http://my.company.com/git/ljpapi"),
                 new File("src/com/company/ljp/Alignment.java", "Alignment.java"),
                 new Branch("feature/feature42D"),
-                32,
+                new LineSelection(32),
                 "http://my.company.com/blob/ljpapi.git/feature!feature42D/src!com!company!ljp!Alignment.java#L32"
             },
         };