+++ tests/unit/uk/co/ben_gibson/git/link/test/Url/Factory/GitHubUrlFactoryTest.java
@@ -1,23 +1,17 @@
 package uk.co.ben_gibson.git.link.test.Url.Factory;
 
 import com.tngtech.java.junit.dataprovider.DataProvider;
-import org.junit.Test;
 import uk.co.ben_gibson.git.link.Git.Branch;
 import uk.co.ben_gibson.git.link.Git.Commit;
 import uk.co.ben_gibson.git.link.Git.Exception.RemoteException;
 import uk.co.ben_gibson.git.link.Git.File;
+import uk.co.ben_gibson.git.link.UI.LineSelection;
 import uk.co.ben_gibson.git.link.Url.Factory.GitHubUrlFactory;
 
 import java.net.MalformedURLException;
 
 public class GitHubUrlFactoryTest extends UrlFactoryTest
 {
-    @Test
-    public void testCanDetermineIfFileAtCommitIsSupported()
-    {
-        assertTrue(new GitHubUrlFactory().canOpenFileAtCommit());
-    }
-
 
     @DataProvider
     public static Object[][] commitProvider() throws MalformedURLException, RemoteException
@@ -54,8 +48,16 @@ public class GitHubUrlFactoryTest extends UrlFactoryTest
                 mockRemote("http://github.com/foo/bar"),
                 new File("src/Bar.java", "Bar.java"),
                 new Commit("f7c244eeea9f8e4ebbeabc1500b90e656f5d0328"),
-                10,
+                new LineSelection(10),
                 "http://github.com/foo/bar/blob/f7c244eeea9f8e4ebbeabc1500b90e656f5d0328/src/Bar.java#L10"
+            },
+            {
+                new GitHubUrlFactory(),
+                mockRemote("http://github.com/foo/bar"),
+                new File("src/Bar.java", "Bar.java"),
+                new Commit("f7c244eeea9f8e4ebbeabc1500b90e656f5d0328"),
+                new LineSelection(10, 20),
+                "http://github.com/foo/bar/blob/f7c244eeea9f8e4ebbeabc1500b90e656f5d0328/src/Bar.java#L10-L20"
             },
         };
     }
@@ -70,8 +72,16 @@ public class GitHubUrlFactoryTest extends UrlFactoryTest
                 mockRemote("https://github.com/foo/bar"),
                 new File("src/Bar.java", "Bar.java"),
                 Branch.master(),
-                10,
+                new LineSelection(10),
                 "https://github.com/foo/bar/blob/master/src/Bar.java#L10"
+            },
+            {
+                new GitHubUrlFactory(),
+                mockRemote("https://github.com/foo/bar"),
+                new File("src/Bar.java", "Bar.java"),
+                Branch.master(),
+                new LineSelection(10, 20),
+                "https://github.com/foo/bar/blob/master/src/Bar.java#L10-L20"
             },
             {
                 new GitHubUrlFactory(),
@@ -94,7 +104,7 @@ public class GitHubUrlFactoryTest extends UrlFactoryTest
                 mockRemote("https://github.com/foo bar/baz"),
                 new File("需求0920-2017/0928.sql", "0928.sql"),
                 Branch.master(),
-                15,
+                new LineSelection(15),
                 "https://github.com/foo%20bar/baz/blob/master/需求0920-2017/0928.sql#L15"
             },
         };