+++ tests/unit/uk/co/ben_gibson/git/link/test/Url/Factory/GogsUrlFactoryTest.java
@@ -1,24 +1,17 @@
 package uk.co.ben_gibson.git.link.test.Url.Factory;
 
 import com.tngtech.java.junit.dataprovider.DataProvider;
-import org.junit.Test;
 import uk.co.ben_gibson.git.link.Git.Branch;
 import uk.co.ben_gibson.git.link.Git.Commit;
 import uk.co.ben_gibson.git.link.Git.Exception.RemoteException;
 import uk.co.ben_gibson.git.link.Git.File;
+import uk.co.ben_gibson.git.link.UI.LineSelection;
 import uk.co.ben_gibson.git.link.Url.Factory.GogsUrlFactory;
 
 import java.net.MalformedURLException;
 
 public class GogsUrlFactoryTest extends UrlFactoryTest
 {
-    @Test
-    public void testCanDetermineIfFileAtCommitIsSupported()
-    {
-        assertTrue(new GogsUrlFactory().canOpenFileAtCommit());
-    }
-
-
     @DataProvider
     public static Object[][] commitProvider() throws MalformedURLException, RemoteException
     {
@@ -54,8 +47,16 @@ public class GogsUrlFactoryTest extends UrlFactoryTest
                 mockRemote("https://try.gogs.io/foo/bar"),
                 new File("src/Bar.java", "Bar.java"),
                 new Commit("f7c244eeea9f8e4ebbeabc1500b90e656f5d0328"),
-                10,
+                new LineSelection(10),
                 "https://try.gogs.io/foo/bar/src/f7c244eeea9f8e4ebbeabc1500b90e656f5d0328/src/Bar.java#L10"
+            },
+            {
+                new GogsUrlFactory(),
+                mockRemote("https://try.gogs.io/foo/bar"),
+                new File("src/Bar.java", "Bar.java"),
+                new Commit("f7c244eeea9f8e4ebbeabc1500b90e656f5d0328"),
+                new LineSelection(10, 20),
+                "https://try.gogs.io/foo/bar/src/f7c244eeea9f8e4ebbeabc1500b90e656f5d0328/src/Bar.java#L10-L20"
             },
         };
     }
@@ -70,8 +71,16 @@ public class GogsUrlFactoryTest extends UrlFactoryTest
                 mockRemote("https://try.gogs.io/foo/bar"),
                 new File("src/Bar.java", "Bar.java"),
                 Branch.master(),
-                10,
+                new LineSelection(10),
                 "https://try.gogs.io/foo/bar/src/master/src/Bar.java#L10"
+            },
+            {
+                new GogsUrlFactory(),
+                mockRemote("https://try.gogs.io/foo/bar"),
+                new File("src/Bar.java", "Bar.java"),
+                Branch.master(),
+                new LineSelection(10, 20),
+                "https://try.gogs.io/foo/bar/src/master/src/Bar.java#L10-L20"
             },
             {
                 new GogsUrlFactory(),
@@ -94,7 +103,7 @@ public class GogsUrlFactoryTest extends UrlFactoryTest
                 mockRemote("https://try.gogs.io/foo bar/baz"),
                 new File("需求0920-2017/0928.sql", "0928.sql"),
                 Branch.master(),
-                15,
+                new LineSelection(15),
                 "https://try.gogs.io/foo%20bar/baz/src/master/需求0920-2017/0928.sql#L15"
             },
         };